Time: 5 ms (58.4%), Space: 16.6 MB (15.43%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-         int maxi = INT_MIN;
-        int prod=1;
+        double maxi = INT_MIN;
+        double prod=1;
 
         for(int i=0;i<nums.size();i++)
         {
